move search type to settings
